Memperbaiki struktur table, memperbaiki fitur CRUD CPL, menambahkan fitur authemtication

+++ database/migrations/2024_10_07_072954_create_cpmks_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->text('deskripsi');
-            $table->integer('skor_maks');
+            $table->float('skor_maks');
             $table->timestamps();
         });
     }
